requirements.txt
Flask==1.1.1
Werkzeug==0.16.0
mysql-connector-python==8.0.19
https://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.0.4.zip#md5=3df394d89300db95163f17c843ef49df